Moved the examples to an example package
